fix(firmware): keep capability writemask writable
