+++ internal/firmware/codegen/codegen_test.go
@@ -27,113 +27,6 @@ func TestGenerateConfigSpaceCOE(t *testing.T) {
 	}
 }
 
-func TestGenerateWritemaskCOE(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU16(0x00, 0x8086)
-	cs.WriteU32(0x10, 0xFFFFF000) // BAR0: 4KB 32-bit memory BAR (scrubbed; dynamic size from CappedBAR0Size)
-
-	wm := GenerateWritemaskCOE(cs)
-
-	if !strings.Contains(wm, "memory_initialization_radix=16") {
-		t.Error("writemask COE should contain radix")
-	}
-
-	// parse data lines into DWORD array
-	lines := strings.Split(wm, "\n")
-	dwords := make([]string, 0, len(lines))
-	for _, l := range lines {
-		l = strings.TrimSpace(l)
-		if l == "" || strings.HasPrefix(l, ";") || strings.HasPrefix(l, "memory") {
-			continue
-		}
-		l = strings.TrimSuffix(l, ",")
-		l = strings.TrimSuffix(l, ";")
-		dwords = append(dwords, l)
-	}
-
-	if len(dwords) != 1024 {
-		t.Fatalf("writemask should have 1024 DWORDs, got %d", len(dwords))
-	}
-
-	// DWORD 0 (VID:DID) must be read-only
-	if dwords[0] != "00000000" {
-		t.Errorf("VID:DID mask should be 00000000, got %s", dwords[0])
-	}
-
-	// DWORD 2 (Rev:ClassCode) must be read-only
-	if dwords[2] != "00000000" {
-		t.Errorf("Rev:ClassCode mask should be 00000000, got %s", dwords[2])
-	}
-
-	// DWORD 4 (BAR0) must match scrubbed BAR size mask (0xFFFFF000)
-	if dwords[4] != "fffff000" {
-		t.Errorf("BAR0 mask should be fffff000 (4KB size e.g.), got %s", dwords[4])
-	}
-
-	// DWORD 5 (BAR1) must be 0 (unused after scrubber clamp)
-	if dwords[5] != "00000000" {
-		t.Errorf("BAR1 mask should be 00000000 (unused), got %s", dwords[5])
-	}
-
-	// DWORD 11 (SubsysIDs) must be read-only
-	if dwords[11] != "00000000" {
-		t.Errorf("SubsysIDs mask should be 00000000, got %s", dwords[11])
-	}
-
-	// DWORD 15 (IntLine/IntPin) - only IntLine writable
-	if dwords[15] != "000000ff" {
-		t.Errorf("IntLine/IntPin mask should be 000000ff, got %s", dwords[15])
-	}
-
-	for i := 0x40 / 4; i < 0x100/4; i++ {
-		if dwords[i] != "00000000" {
-			t.Errorf("cap region DWORD[%d] should be 00000000, got %s", i, dwords[i])
-		}
-	}
-
-	// extended config space (0x100+) should be read-only
-	for i := 0x100 / 4; i < len(dwords); i++ {
-		if dwords[i] != "00000000" {
-			t.Errorf("ext cap DWORD[%d] should be 00000000, got %s", i, dwords[i])
-			break
-		}
-	}
-}
-
-func TestGenerateWritemaskCOE_KnownCapabilities(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU16(0x06, 0x0010)
-	cs.WriteU8(0x34, 0x40)
-
-	cs.WriteU8(0x40, pci.CapIDPowerManagement)
-	cs.WriteU8(0x41, 0x50)
-	cs.WriteU8(0x50, pci.CapIDMSIX)
-	cs.WriteU8(0x51, 0x60)
-	cs.WriteU8(0x60, pci.CapIDPCIExpress)
-	cs.WriteU8(0x61, 0x00)
-
-	wm := GenerateWritemaskCOE(cs)
-	dwords := parseCOEDwords(t, wm)
-
-	if dwords[16] != "00000000" {
-		t.Errorf("PM header DWORD should be read-only, got %s", dwords[16])
-	}
-	if dwords[17] != "00000103" {
-		t.Errorf("PMCSR DWORD should be 00000103, got %s", dwords[17])
-	}
-	if dwords[20] != "c0000000" {
-		t.Errorf("MSI-X control DWORD should expose only enable/mask bits, got %s", dwords[20])
-	}
-	if dwords[24] != "00000000" {
-		t.Errorf("PCIe header DWORD should be read-only, got %s", dwords[24])
-	}
-	if dwords[26] != "0000ffff" {
-		t.Errorf("PCIe Device Control/Status DWORD should be writable, got %s", dwords[26])
-	}
-}
-
 func TestGenerateBarContentCOE_Empty(t *testing.T) {
 	coe := GenerateBarContentCOE(nil, 4096)
 	if !strings.Contains(coe, "Zero-filled") {
@@ -212,156 +105,6 @@ func TestGenerateMSIXTableHex_Empty(t *testing.T) {
 	}
 }
 
-// writemask for a donor with BAR0=0 that was scrubbed to 4KB memory BAR (default).
-// this simulates the exact scenario where clampBARsToFPGA creates BAR0 (size from ctx/Capped).
-func TestWritemask_ScrubbedBAR0FromZero(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU16(0x00, 0x1102) // Creative Labs
-	cs.WriteU16(0x02, 0x0012)
-	cs.WriteU32(0x10, 0xFFFFF000) // BAR0 = 4KB mem (from scrubber)
-
-	wm := GenerateWritemaskCOE(cs)
-	dwords := parseCOEDwords(t, wm)
-
-	// BAR0 writemask should allow sizing writes
-	if dwords[4] != "fffff000" {
-		t.Errorf("BAR0 mask = %s, want fffff000 (4KB size mask e.g.)", dwords[4])
-	}
-}
-
-// simulate BAR sizing: host writes 0xFFFFFFFF, reads back, decodes size.
-func TestWritemask_BARSizingSimulation(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU32(0x10, 0xFFFFF000) // 4KB (default) 32-bit memory BAR
-
-	wm := GenerateWritemaskCOE(cs)
-	dwords := parseCOEDwords(t, wm)
-
-	// writemask must be 0xFFFFF000 for BAR0
-	if dwords[4] != "fffff000" {
-		t.Fatalf("BAR0 mask = %s, want fffff000", dwords[4])
-	}
-
-	// simulate host BAR sizing:
-	// 1. host writes 0xFFFFFFFF to BAR0
-	// 2. BRAM stores: 0xFFFFFFFF & writemask = 0xFFFFF000
-	// 3. host reads back 0xFFFFF000
-	// 4. host decodes: ~(0xFFFFF000 & ~0xF) + 1 = ~0xFFFFF000 + 1 = 0x1000 = 4KB
-	hostWrite := uint32(0xFFFFFFFF)
-	mask := uint32(0xFFFFF000)
-	bramResult := hostWrite & mask
-	if bramResult != 0xFFFFF000 {
-		t.Errorf("BAR sizing result = 0x%08X, want 0xFFFFF000", bramResult)
-	}
-	sizeBits := bramResult & ^uint32(0x0F)
-	barSize := ^sizeBits + 1
-	if barSize != 4096 {
-		t.Errorf("decoded BAR size = %d, want 4096 (for this 4k mask test case)", barSize)
-	}
-}
-
-// writemask with IO BAR donor that was overwritten to memory BAR by scrubber.
-func TestWritemask_IOBAROverwrittenToMemory(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU32(0x10, 0xFFFFF000) // scrubber forced memory BAR
-
-	wm := GenerateWritemaskCOE(cs)
-	dwords := parseCOEDwords(t, wm)
-
-	// must be memory BAR mask, NOT IO mask, NOT zero
-	if dwords[4] == "00000000" {
-		t.Error("BAR0 mask should not be zero after scrubber creates memory BAR")
-	}
-	if dwords[4] != "fffff000" {
-		t.Errorf("BAR0 mask = %s, want fffff000", dwords[4])
-	}
-}
-
-// identity registers must be read-only in writemask.
-func TestWritemask_IdentityRegistersReadOnly(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU16(0x00, 0x1102)     // VID
-	cs.WriteU16(0x02, 0x0012)     // DID
-	cs.WriteU32(0x08, 0x04030000) // ClassCode + RevID
-	cs.WriteU16(0x2C, 0x1102)     // SubsysVID
-	cs.WriteU16(0x2E, 0x0012)     // SubsysDID
-	cs.WriteU32(0x10, 0xFFFFF000) // BAR0
-
-	wm := GenerateWritemaskCOE(cs)
-	dwords := parseCOEDwords(t, wm)
-
-	roRegs := map[int]string{
-		0:  "VID:DID",
-		2:  "RevisionID:ClassCode",
-		10: "CardBus CIS",
-		11: "SubsysVID:SubsysDID",
-		12: "Expansion ROM",
-		13: "CapPtr",
-		14: "Reserved",
-	}
-	for dw, name := range roRegs {
-		if dwords[dw] != "00000000" {
-			t.Errorf("%s (DWORD %d) mask = %s, want 00000000 (read-only)", name, dw, dwords[dw])
-		}
-	}
-}
-
-// command/status register must be fully writable.
-func TestWritemask_CommandStatusWritable(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU32(0x10, 0xFFFFF000)
-
-	wm := GenerateWritemaskCOE(cs)
-	dwords := parseCOEDwords(t, wm)
-
-	if dwords[1] != "ffffffff" {
-		t.Errorf("Command:Status mask = %s, want ffffffff", dwords[1])
-	}
-}
-
-// prefetchable memory BAR: type bits [3:0] must be read-only.
-func TestWritemask_PrefetchableBAR(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU32(0x10, 0xFFFFF008) // 4KB (default) prefetchable 32-bit memory
-
-	wm := GenerateWritemaskCOE(cs)
-	dwords := parseCOEDwords(t, wm)
-
-	// mask should be 0xFFFFF008 & 0xFFFFFFF0 = 0xFFFFF000
-	// type bits [3:0] read-only, size bits writable
-	if dwords[4] != "fffff000" {
-		t.Errorf("prefetchable BAR mask = %s, want fffff000", dwords[4])
-	}
-}
-
-// multiple BARs: each should get its own mask.
-func TestWritemask_MultipleBARs(t *testing.T) {
-	cs := pci.NewConfigSpace()
-	cs.Size = pci.ConfigSpaceSize
-	cs.WriteU32(0x10, 0xFFFFF000) // BAR0: 4KB memory
-	cs.WriteU32(0x14, 0x00000000) // BAR1: unused
-	cs.WriteU32(0x18, 0xFFFF0000) // BAR2: 64KB memory
-
-	wm := GenerateWritemaskCOE(cs)
-	dwords := parseCOEDwords(t, wm)
-
-	if dwords[4] != "fffff000" {
-		t.Errorf("BAR0 mask = %s, want fffff000", dwords[4])
-	}
-	if dwords[5] != "00000000" {
-		t.Errorf("BAR1 mask = %s, want 00000000 (unused)", dwords[5])
-	}
-	if dwords[6] != "ffff0000" {
-		t.Errorf("BAR2 mask = %s, want ffff0000", dwords[6])
-	}
-}
-
 // config space COE must contain correct DWORDs for a scrubbed config space.
 func TestConfigSpaceCOE_ScrubbedDeviceIdentity(t *testing.T) {
 	cs := pci.NewConfigSpace()
@@ -393,24 +136,4 @@ func TestConfigSpaceHex_DWORDFormat(t *testing.T) {
 	if !strings.Contains(hex, "FFFFF000 // [010]") {
 		t.Error("HEX should contain BAR0 at offset [010]")
 	}
-}
-
-// parseCOEDwords extracts hex DWORD strings from a COE file.
-func parseCOEDwords(t *testing.T, coe string) []string {
-	t.Helper()
-	lines := strings.Split(coe, "\n")
-	dwords := make([]string, 0, len(lines))
-	for _, l := range lines {
-		l = strings.TrimSpace(l)
-		if l == "" || strings.HasPrefix(l, ";") || strings.HasPrefix(l, "memory") {
-			continue
-		}
-		l = strings.TrimSuffix(l, ",")
-		l = strings.TrimSuffix(l, ";")
-		dwords = append(dwords, l)
-	}
-	if len(dwords) != 1024 {
-		t.Fatalf("expected 1024 DWORDs, got %d", len(dwords))
-	}
-	return dwords
 }
